third_party/WebKit/LayoutTests/fullscreen/compositor-touch-hit-rects-fullscreen-video-controls-expected.txt

This test makes sure that touch hit rects are reported for fullscreen HTML5 video control elements even when there is a document handler.

Should have single rect on document before fullscreen
handler: #document scrolling (0, 0, 800, 600)

EVENT(webkitfullscreenchange)
Should keep rect on document
handler: DIV (0, 0, 800, 600)
handler: DIV (0, 0, 800, 600)
handler: #document scrolling (0, 0, 800, 600)

END OF TEST
